更改场地金额类型

+++ gym_java/src/main/java/edu/gdou/gym_java/service/FieldService.java
@@ -64,11 +64,11 @@ public interface FieldService extends IService<Field> {
 
     Integer queryTypeByName(String typeName);
 
-    Integer queryMoneyByTimeId(Integer timeId);
+    Double queryMoneyByTimeId(Integer timeId);
 
     Boolean checkCancelTime(FieldCheck fieldCheck);
 
-    Integer querySumMoney(String beginDate, String endDate);
+    Double querySumMoney(String beginDate, String endDate);
 
     Boolean deleteType(FieldType fieldType);
